+++ resolwe_bio/tests/processes/test_reads_filtering.py
@@ -1,132 +1,9 @@
 # pylint: disable=missing-docstring
 from resolwe.test import tag_process
-from resolwe_bio.utils.test import BioProcessTestCase, skipDockerFailure
+from resolwe_bio.utils.test import BioProcessTestCase
 
 
 class ReadsFilteringProcessorTestCase(BioProcessTestCase):
-
-    @tag_process('prinseq-lite-single')
-    def test_prinseq_single(self):
-        with self.preparation_stage():
-            reads = self.prepare_reads()
-            inputs = {'reads': reads.pk}
-
-        filtered_reads = self.run_process('prinseq-lite-single', inputs)
-        self.assertFiles(filtered_reads, 'fastq', ['filtered_reads_prinseq_single.fastq.gz'], compression='gzip')
-        del filtered_reads.output['fastqc_url'][0]['total_size']  # Non-deterministic output.
-        self.assertFields(filtered_reads, "fastqc_url", [{'file': 'fastqc/reads_fastqc/fastqc_report.html',
-                                                          'refs': ['fastqc/reads_fastqc'],
-                                                          'size': 314749}])
-
-    @tag_process('prinseq-lite-paired')
-    def test_prinseq_paired(self):
-        with self.preparation_stage():
-            inputs = {
-                'src1': ['rRNA forw.fastq.gz'],
-                'src2': ['rRNA_rew.fastq.gz']}
-            reads = self.run_process('upload-fastq-paired', inputs)
-
-        inputs = {'reads': reads.pk}
-        filtered_reads = self.run_process('prinseq-lite-paired', inputs)
-        self.assertFiles(filtered_reads, 'fastq', ['filtered_reads_prinseq_paired_fw.fastq.gz'], compression='gzip')
-        self.assertFiles(filtered_reads, 'fastq2', ['filtered_reads_prinseq_paired_rw.fastq.gz'], compression='gzip')
-        del filtered_reads.output['fastqc_url'][0]['total_size']  # Non-deterministic output.
-        self.assertFields(filtered_reads, "fastqc_url", [{'file': 'fastqc/rRNA forw_fastqc/fastqc_report.html',
-                                                          'refs': ['fastqc/rRNA forw_fastqc'],
-                                                          'size': 347773}])
-        del filtered_reads.output['fastqc_url2'][0]['total_size']  # Non-deterministic output.
-        self.assertFields(filtered_reads, "fastqc_url2", [{'file': 'fastqc/rRNA_rew_fastqc/fastqc_report.html',
-                                                           'refs': ['fastqc/rRNA_rew_fastqc'],
-                                                           'size': 340697}])
-
-    @tag_process('fastq-mcf-single')
-    def test_fastqmcf_single(self):
-        with self.preparation_stage():
-            reads = self.prepare_reads()
-
-        inputs = {'reads': reads.pk}
-        filtered_reads = self.run_process('fastq-mcf-single', inputs)
-
-        self.assertFiles(filtered_reads, 'fastq', ['filtered_reads_fastqmcf_single.fastq.gz'], compression='gzip')
-        del filtered_reads.output['fastqc_url'][0]['total_size']  # Non-deterministic output.
-        self.assertFields(filtered_reads, "fastqc_url", [{'file': 'fastqc/reads_fastqc/fastqc_report.html',
-                                                          'refs': ['fastqc/reads_fastqc'],
-                                                          'size': 305101}])
-
-    @tag_process('fastq-mcf-paired')
-    def test_fastqmcf_paired(self):
-        with self.preparation_stage():
-            inputs = {
-                'src1': ['rRNA forw.fastq.gz'],
-                'src2': ['rRNA_rew.fastq.gz']}
-            reads = self.run_process('upload-fastq-paired', inputs)
-
-        inputs = {'reads': reads.pk}
-        filtered_reads = self.run_process('fastq-mcf-paired', inputs)
-        self.assertFiles(filtered_reads, 'fastq', ['filtered_reads_fastqmcf_paired_fw.fastq.gz'], compression='gzip')
-        self.assertFiles(filtered_reads, 'fastq2', ['filtered_reads_fastqmcf_paired_rw.fastq.gz'], compression='gzip')
-        del filtered_reads.output['fastqc_url'][0]['total_size']  # Non-deterministic output.
-        self.assertFields(filtered_reads, "fastqc_url", [{'file': 'fastqc/rRNA forw_fastqc/fastqc_report.html',
-                                                          'refs': ['fastqc/rRNA forw_fastqc'],
-                                                          'size': 347791}])
-        del filtered_reads.output['fastqc_url2'][0]['total_size']  # Non-deterministic output.
-        self.assertFields(filtered_reads, "fastqc_url2", [{'file': 'fastqc/rRNA_rew_fastqc/fastqc_report.html',
-                                                           'refs': ['fastqc/rRNA_rew_fastqc'],
-                                                           'size': 340715}])
-
-    @skipDockerFailure("Skip until Docker image with iCount is supported on Travis.")
-    @tag_process('sortmerna-single')
-    def test_sortmerna_single(self):
-        with self.preparation_stage():
-            reads = self.prepare_reads(['rRNA forw.fastq.gz'])
-            rrnadb_1 = self.run_process('upload-fasta-nucl', {'src': 'silva-arc-16s-id95.fasta.gz'})
-            rrnadb_2 = self.run_process('upload-fasta-nucl', {'src': 'silva-arc-23s-id98.fasta.gz'})
-
-        inputs = {
-            'reads': reads.id,
-            'database_selection': [rrnadb_1.id, rrnadb_2.id],
-            'options': {'threads': 2, 'sam': True}}
-        filtered_reads = self.run_process('sortmerna-single', inputs)
-        self.assertFiles(filtered_reads, 'fastq', ['reads_wo_rRNA_single.fastq.gz'], compression='gzip')
-        self.assertFile(filtered_reads, 'fastq_rRNA', 'reads_rRNA_single.fastq.gz', compression='gzip')
-        self.assertFields(filtered_reads, 'fastq_rRNA_sam', {'file': 'rRNA forw_rRNA.sam'})
-        self.assertFields(filtered_reads, 'stats', {'file': 'stats.log'})
-        self.assertFields(filtered_reads, "fastqc_url",
-                          [{'file': 'fastqc/rRNA forw_filtered_fastqc/fastqc_report.html',
-                            'refs': ['fastqc/rRNA forw_filtered_fastqc'],
-                            'size': 345492}])
-
-    @skipDockerFailure("Skip until Docker image with iCount is supported on Travis.")
-    @tag_process('sortmerna-paired')
-    def test_sortmerna_paired(self):
-        with self.preparation_stage():
-            inputs = {
-                'src1': ['rRNA forw.fastq.gz'],
-                'src2': ['rRNA_rew.fastq.gz']}
-            reads = self.run_process('upload-fastq-paired', inputs)
-
-            rrnadb_1 = self.run_process('upload-fasta-nucl', {'src': 'silva-arc-16s-id95.fasta.gz'})
-            rrnadb_2 = self.run_process('upload-fasta-nucl', {'src': 'silva-arc-23s-id98.fasta.gz'})
-
-        inputs = {
-            'reads': reads.id,
-            'database_selection': [rrnadb_1.id, rrnadb_2.id],
-            'options': {'threads': 2, 'sam': True}}
-
-        filtered_reads = self.run_process('sortmerna-paired', inputs)
-        self.assertFiles(filtered_reads, 'fastq', ['reads_wo_rRNA_paired_forw.fastq.gz'], compression='gzip')
-        self.assertFiles(filtered_reads, 'fastq2', ['reads_wo_rRNA_paired_rew.fastq.gz'], compression='gzip')
-        self.assertFile(filtered_reads, 'fastq_rRNA', 'reads_rRNA_paired.fastq.gz', compression='gzip')
-        self.assertFields(filtered_reads, 'fastq_rRNA_sam', {'file': 'rRNA forw_rRNA.sam'})
-        self.assertFields(filtered_reads, 'stats', {'file': 'stats.log'})
-        self.assertFields(filtered_reads, "fastqc_url",
-                          [{'file': 'fastqc/rRNA forw_filtered_fastqc/fastqc_report.html',
-                            'refs': ['fastqc/rRNA forw_filtered_fastqc'],
-                            'size': 345492}])
-        self.assertFields(filtered_reads, "fastqc_url2",
-                          [{'file': 'fastqc/rRNA_rew_filtered_fastqc/fastqc_report.html',
-                            'refs': ['fastqc/rRNA_rew_filtered_fastqc'],
-                            'size': 347212}])
 
     @tag_process('trimmomatic-single')
     def test_trimmomatic_single(self):
@@ -188,26 +65,6 @@ class ReadsFilteringProcessorTestCase(BioProcessTestCase):
         del filtered_reads.output['fastqc_url2'][0]['total_size']  # Non-deterministic output.
         self.assertFields(filtered_reads, "fastqc_url2", [{'file': 'fastqc/rRNA_rew_fastqc/fastqc_report.html',
                                                            'refs': ['fastqc/rRNA_rew_fastqc']}])
-
-    @tag_process('hsqutils-trim')
-    def test_hsqutils_trim(self):
-        with self.preparation_stage():
-            inputs = {
-                'src1': ['hsqutils_reads_mate1_paired_filtered.fastq.gz'],
-                'src2': ['hsqutils_reads_mate2_paired_filtered.fastq.gz']}
-            reads = self.run_processor('upload-fastq-paired', inputs)
-
-            probe = self.run_processor('upload-file', {'src': 'hsqutils_probe_info.txt'})
-
-        inputs = {'reads': reads.id,
-                  'probe': probe.id}
-
-        hsqutils_trimm = self.run_processor('hsqutils-trim', inputs)
-
-        self.assertFiles(hsqutils_trimm, 'fastq', ['hsqutils_reads_trimmed_mate1.fastq.gz'],
-                         compression='gzip')
-        self.assertFiles(hsqutils_trimm, 'fastq2', ['hsqutils_reads_trimmed_mate2.fastq.gz'],
-                         compression='gzip')
 
     @tag_process('cutadapt-single')
     def test_cutadapt_single(self):